+++ changelog.md
@@ -1,6 +1,6 @@
-# Version 2025.2.2 (2025-02-02)
+# Version 2025.2.2 (2025-02-03)
 
--
+- Ignore devices with unknown battery
 
 # Version 2025.2.1 (2025-02-02)